Rename advanced deeplink app and package
Rename second app to advanceddeeplinkapp and reuse the same deeplink.advanced package
