+++ advanceddeeplinkapp/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:theme="@style/Theme.Nav3Recipes"
         android:enableOnBackInvokedCallback="true">
         <activity
-            android:name="MainActivity"
+            android:name="AdvancedDeeplinkAppActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:launchMode="singleTop"
