+++ advanceddeeplinkapp/src/main/res/drawable/ic_launcher_foreground.xml
