+++ advanceddeeplinkapp/src/main/res/values/colors.xml
