+++ advanceddeeplinkapp/src/main/res/values/strings.xml
