+++ app/src/main/AndroidManifest.xml
@@ -180,7 +180,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".deeplink.advanced.AdvancedCreateAdvancedDeepLinkActivity"
+            android:name=".deeplink.advanced.AdvancedCreateDeepLinkActivity"
             android:exported="true"
             android:theme="@style/Theme.Nav3Recipes">
         </activity>
