+++ app/src/main/java/com/example/nav3recipes/deeplink/basic/README.md
@@ -19,5 +19,5 @@ See `MainActivity.deepLinkPatterns` for the actual url pattern of each.
 
 This recipe consists of three main packages:
 1. `basic.deeplink` - Contains the two activities
-2. `basic.deeplink.ui` - Contains the activity UI code, i.e. i.e. global string variables, deeplink URLs etc
+2. `basic.deeplink.ui` - Contains the activity UI code, i.e. global string variables, deeplink URLs etc
 3. `basic.deeplink.util` - Contains the classes and helper methods to parse and match the deeplinks
